+++ src/main/java/roomescape/dto/MemberRequest.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.dto;
 
 public record MemberRequest(String name, String email, String password) {
 }
