+++ src/main/java/roomescape/infra/auth/AuthenticationPrincipalArgumentResolver.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.infra.auth;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
