+++ src/main/java/roomescape/infra/auth/CheckAdminInterceptor.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.infra.auth;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
@@ -18,6 +18,7 @@ public class CheckAdminInterceptor implements HandlerInterceptor {
 
     @Override
     public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler) {
+        // role이랑 이름 담아서 보내면 된다.
         Cookie[] cookies = request.getCookies();
         if (cookies == null) {
             response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
